backend/requirements.txt
beautifulsoup4
requests
requests-html
lxml[html_clean]
python-dotenv
lxml_html_clean
pyppeteer
flask
flask-cors
streamlit
transformers 
pipeline
torch
sentencepiece
datasets[audio]
soundfile
accelerate
sentence-transformers
google-generativeai
gunicorn>=21.2